+++ agent/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## 0.1.0-dev
+
+* Allow to send audio to other endpoints [#68](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/68)