+++ ex_webrtc/CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix track_types count when subscription fails [#63](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/63)
 * Update protos [#64](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/64)
 * Allow ignoring specified endpoints [#66](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/66)
+* Dependency to protobuf main branch [#68](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/68)
 
 ## 0.1.0
 * Initial release [#406](https://github.com/fishjam-dev/membrane_rtc_engine/pull/406)